requirements-api.txt
flask==2.0.1
flask-cors==3.0.10
numpy>=1.20.0
scikit-image>=0.17.2
pyyaml>=5.3
tensorflow>=2.0.0
pillow>=8.0.1
gunicorn>=20.1.0
requests>=2.28.1 